+++ src/CallWebhookJob.php
@@ -7,6 +7,7 @@ use GuzzleHttp\Client;
 use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Psr7\Response;
+use GuzzleHttp\TransferStats;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
@@ -53,6 +54,8 @@ class CallWebhookJob implements ShouldQueue
 
     private ?string $errorMessage = null;
 
+    private ?TransferStats $transferStats = null;
+
     public function handle()
     {
         /** @var \GuzzleHttp\Client $client */
@@ -69,6 +72,9 @@ class CallWebhookJob implements ShouldQueue
                 'timeout' => $this->requestTimeout,
                 'verify' => $this->verifySsl,
                 'headers' => $this->headers,
+                'on_stats' => function (TransferStats $stats) {
+                    $this->transferStats = $stats;
+                },
             ], $body));
 
             if (! Str::startsWith($this->response->getStatusCode(), 2)) {
@@ -132,7 +138,8 @@ class CallWebhookJob implements ShouldQueue
             $this->response,
             $this->errorType,
             $this->errorMessage,
-            $this->uuid
+            $this->uuid,
+            $this->transferStats
         ));
     }
 }