+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to `laravel-webhook-server` will be documented in this file
 
+## 1.13.0 - 2021-04-28
+
+- add `dispatchSync`
+
+## 1.12.0 - 2021-04-20
+
+- pass Guzzle TransferStats into resulting Event (#81)
+
+## 1.11.3 - 2021-04-02
+
+- fix for missing default headers when using withHeaders (#79)
+
+## 1.11.2 - 2021-03-17
+
+- dispatch should return the PendingDispatch (#74)
+
 ## 1.11.1 - 2020-12-15
 
 - fix exception name for invalid signers (#67)